[turboshaft] Introduce Smi Constants
Currently, constant Smis are represented in Turboshaft as
TaggedBitcast(Constant(kind = Word32/Word64). This causes the
Instruction Selector to later allocate a register for such Smis,
instead of using them as Constants operands.

This CL introduces a Smi kind for ConstantOp, which in addition for
fixing this Instruction Selector problem, will also be useful to more
easily pattern match Smi constant, and will be useful when working on
the front-end, where quite a lot of constant Smis will be created.

Bug: v8:12783
Change-Id: I0f2aa1b03ab512b111e65e4e20f84b7afc23813a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4953364
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92221}