[turboshaft][wasm] RecreateSchedule: Mark exception block deferred
It was noted that RecreateSchedule does not preserve the deferredness
of catch blocks ("IfException" in Turbofan).
Note that this was noted because with deferred block the
RegisterAllocator is significantly slower in certain situations, so if
this CL causes significant compile time regressions, these should be
investigated in the RegisterAllocator.

Bug: chromium:1519099, v8:12783, v8:14108
Change-Id: I24cb0fd79d950a7c583a59928ac0d80b8eb27f57
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5257662
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92141}