Introduce a SmiCompressionScheme
In addition to the V8HeapCompressionScheme,
TrustedSpaceCompressionScheme, and ExternalCodeCompressionScheme, this
CL introduces a SmiCompressionScheme which can be used in places where
we know that all objects we deal with will be Smis (e.g. for Smi-field
accessors).

This prepares a CL to introduce a ProtectedFixedArray, which will
require specifying the CompressionScheme for any Shape for
TaggedArrayBase. In the case of Smis, there is no point in specifying
V8HeapCompressionScheme or any of the other schemes.

We also add some more "V8_ASSUME" and "constexpr" to maybe get some
better code generated by clang.

R=ishell@chromium.org

Bug: v8:14564
Change-Id: Ia6b860b168f87f5064380bfa86cf513ffa4f3a49
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5252525
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92145}