[heap] Check that we do not allocate in fast c calls
Fast c calls are not allowed to allocate or start a GC. Add DCHECKs
and CHECKs to ensure that this is indeed the case through all of V8.

Wasm-c-api was using fast c calls with allocation as well. We could
either disable the checks for it or simply do not set
fast_c_call_caller_fp for those calls. I chose the latter approach
because then we can ensure the (D)CHECKs hold all the time.

Bug: chromium:1523607
Change-Id: Iebc0ee3eae791ed2ae2435d2ca6f1d34937b3476
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5253159
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92153}