Fix clearing of dead TRUSTED_TO_TRUSTED remembered set entries
During mark-compact GC, it can happen that an object with slots recorded
in in a remembered set is freed (e.g. because there is custom weak
marking logic for the object, as is the case for e.g. baseline Code). In
that case, the remembered set entries must be cleared prior to
potentially evacuating other objects into the now-free memory. This is
done in Sweeper::CleanupRememberedSetEntriesForFreedMemory, which was,
however, missing the TRUSTED_TO_TRUSTED set. This CL fixes that.

Bug: chromium:1524115
Change-Id: I945284a636b4cd2ab4d3f198544ac41f1c1f7695
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5258483
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Auto-Submit: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92164}