[liftoff][arm][arm64] Expand macros for binops
Reduce complexity of the LiftoffAssembler by removing the macros for
binary (and some unary) operations.
The multiplication instruction can only operate on registers, so it does
not fit the existing macros, and before adding a new variant it's way
simpler to just define each operation in isolation.

R=ahaas@chromium.org

Bug: v8:14564
Change-Id: I91db353c1d5b16e25cb9d7029970dc3f16752d70
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5268504
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92200}