[handles] Process non-nestable tasks with trivial stack in samples/shell
When pumping the message loop, we may trigger the execution of a
non-nestable task (e.g., a stackless GC), which assumes that there are
no handles on the current stack. With v8_enable_direct_local=true
(which requires conservative stack scanning), violating this invariant
may result in a crash. This CL fixes the V8 shell provided as a sample
to correctly pump the message loop.

Bug: v8:13270
Change-Id: I5ed594f5d1b3a0d4514bf2d36f21789f3942fa4b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5262697
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92309}