[inspector] Don't interrupt while building console messages
This CL fixes a use-after-free of the InspectorSession. This can
happen for workers since they are killed (and the DevToolsSession
detached) during an interrupt.

The problem is that we might have already entered the inspector and
the inspector entered JavaScript land. This could mean that we detach
a session on which we are currently operating on (e.g. building a
console message to send it to the frontend).

The fix is to postpone interrupts until after we are done building
console messages so we don't lose the DevTools session half-way
through it.

R=dsv@chromium.org

Fixed: b:323813642
Change-Id: I495d926830bc0ed129b0632d454b2d94f3123180
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5272444
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Eric Leese <leese@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92226}