[regalloc] Fix quadratic behavior with interval cache
When a LiveRange does not cover the current position,
{LiveRange::Covers()} does not advance the {current_interval_} cache.
This is because the last processed interval in the loop starts after
the current position.
This can result in quadratic behavior where the cache never moves and we
iterate an increasingly longer list of intervals.

The fix is to pass the last processed interval *starting before the
current position* to {AdvanceLastProcessedMarker()}.

R=dlehmann@chromium.org

Bug: b/41496962
Change-Id: I048656d7f52352d5985927808c0771e8134a5928
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5272824
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92232}