[turboshaft][wasm] WasmGCTypeReducer: Improve tracking of unreachable blocks
This CL addresses the known cases where unreachability did not
propagate to succeeding blocks.
this change also adds DCHECKs that we only see the bottom type as a
result of unreachable predecessors.
Still, as the bottom type handling is still in the regular code,
if those DCHECKs are hit, it shouldn't be an observable bug in any
release builds.

Bug: v8:14108
Change-Id: I715c1e2fa8cbd7bce1edbac6ced61cc75367e9c6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5277238
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92231}