[d8][handles] Process non-nestable tasks with trivial stack
When pumping the message loop, we may trigger the execution of a
non-nestable task (e.g., a stackless GC), which assumes that there are
no handles on the current stack. With v8_enable_direct_local=true
(which requires conservative stack scanning), violating this invariant
may result in a crash. This CL fixes d8 to correctly pump the message
loop.

Bug: v8:13270
Change-Id: Ie42ac21719572c1fb533a5f084a271593d46757f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5258097
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92307}