[maglev] Negative index does not invalidate NoElements protector
When alllowing OOB access, we rely on no elements protector to
invalidate the optimization. But since negative index is not
consider an element access, it is transformed to a name and
loaded as a property, it does not invalidate the protector.

Here (as in TF) we force a deoptimization if index is negative.

Bug: v8:7700, chromium:41496337
Change-Id: I1eb7287f7a26e97e44df32cc63cc7aa633608532
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5286015
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92277}