[wasm] Add generic wasm-to-js wrapper for invalid signatures
So far a specific wasm-to-js wrapper got compiled when a wrapper was
requested for a signature that was not compatible with JS, e.g. a
signature that contains a simd. This CL introduces a generic wrapper for
invalid signatures that throws an exception unconditionally.

There was a code path where the generic wrapper was used for invalid
signatures. This code path now also  uses the new wrapper for invalid
signatures.

R=thibaudm@chromium.org

Bug: chromium:323694592
Change-Id: I79fe09c864fef145292fcf2ec428352b6b4f6980
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5285394
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92276}