[wasm] Use correct signature index for tier-up of wasm-to-js wrapper
The wasm-to-js wrapper tierup used the canonicalized signature id lookup
for module-independent signatures to look up the canonicalized signature
id of module-specific signatures. With this CL the signature id is
looked up with the function index of imported functions and from the
dispatch table for indirect function calls instead.

R=jkummerow@chromium.org

Bug: chromium:324596281
Change-Id: I5e47e48ad5e2f103a4c731302ebf3464b59762b8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5291374
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92300}