Use FullPtrComparer in MarkTransitiveClosureLinear
The key_to_values unoredered_multimap is used to check if a newly
discovered, live object is used as key in a WeakMap, in which case the
corresponding value must also be marked now. This map is indexed using
the pointer to the HeapObject, and so it must use full pointer
comparison (rather than compressed pointer comparison) as it may see
pointers to objects in other pointer compression cages.

Bug: b/324656752
Change-Id: Ie97175aa1c0b6bdad012d4537c91ed1a0cbb68e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5291384
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92295}