[compiler] Add fake handles slots to safepoint table
When emitting a safepoint in a fast C call, the indirect handles
passed to the fast callback, need to be included in the safepoint.
This is necessary such that the GC can find those slots and update
pointers in case that object got relocated during GC.

This CL implements this by adding an is_tagged field to a StackSlot.
When set the stack slot should become part of the safepoint.
Currently these slots are simply made part of all safepoints. However,
this requires us to initially clear the slots in the function
prologue. This is necessary because all slots need to contain valid
pointers on a GC.

Bug: chromium:325123663
Change-Id: I5f9aeb507fb3aecf5776ac7ad1a6b1ea75b1d66b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5291382
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92312}