[sandbox] Change the value of kInvalidEntrypointTag
We should not use 0xffff as the invalid code entrypoint tag as this may
cause a pointer to look like a freelist entry if the other 2 top bytes
are also 0xffff (which probably shouldn't happen under normal
circumstances). Instead, this CL now explicitly creates a
kFreeCodePointerTableEntryTag as 0xffff and changes the value of the
kInvalidEntrypointTag to 0xff.

Bug: b/40948502
Change-Id: I3f6784489ae0de7a2869326c7f2e05770dc7500e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5297054
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Auto-Submit: Samuel Groß <saelo@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92337}