[heap] Add DCHECK for Heap::FinishSweepingIfOutOfWork
The stack trace for the crash in [1] shows that
FinishSweepingIfOutOfWork is called, but there is still sweeping
work to do. This CL adds a DCHECK to find similar cases in the
future and let fuzzers find them as well.
(the stack trace in [1] should be resolved by crrev.com/c/5296146)

Drive-by: Remove unused DrainSweepingWorklistForSpace method.

[1] https://crash.corp.google.com/browse?q=ReportID%3D%2739f026b4e907a2be%27&stbtiq=&reportid=&index=0

Bug: b/40943580
Change-Id: Icb2812ba733bcc02c94f0886f189cfa2a0bfd20d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5293616
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92341}