Restore promise silencing behavior
Follow-up to https://crrev.com/c/v8/v8/+/5291617

Previous refactor removed the check for the JSPromise->is_silent
bit when stopping on promise rejections, as it was unused in v8.
However this is part of the v8 public API and is used by
blink streams and could also be used by non-chromium v8 users,
so the debugger may now pause in locations it shouldn't.
This change restores the silencing behavior and adds a test.

Bug: b:40283985, b:40283987, chromium:40150651
Change-Id: Id68a4e619945582166253153fa80d78a2bc64f16
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5300295
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Eric Leese <leese@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92339}