[wasm] Introduce new heap type 'imported string'
This heap type does not exist in the wasm spec. It is used as an
internal type to mark externref values in our optimizations for which
we know that it is a (potentially nullable) JS string.

Previously the wasm string type from the stringref proposal was used
for that. This can cause all kinds of issues as ref.string is a subtype
of any, not extern. (It also uses a different null representation.)

Fixed: chromium:324747822
Change-Id: I9cc2975b8612510de18e02d3b66ed54489d9cc22
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5293801
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92348}