[wasm-compile-fuzzer] Delete recursive groups from type canonicalizer
The wasm compile fuzzer creates random types every run that are saved as recursive groups in the type canonicalizer. It is unnecessary to keep them for the fuzzer runs.
This CL adds a method that deletes the storages used by the type canonicalizer for every fuzzer test run.

Change-Id: I5919e020703795846f210e0012518dc689a0c4be
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5300302
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Eva Herencsárová <evih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92364}