[flags] Add example fuzz test
This adds a test using https://github.com/google/fuzztest/. The test
itself serves more as an example of how to add such a test. We
don't expect it to find much. We might delete it again as soon as
we have enough better examples.

Bugs: 41497291, 40941204
Change-Id: I7f27ba9079fb7958bbbadfcdcea395d4ba9042e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5300802
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92377}