[tagged-ptr] Introduce MaybeWeak for MaybeObject
Introduce MaybeWeak<T> as a weakness sentinel type, representing a value
in a Tagged<>/Handle<>/TNode<> which can be weak. This fulfils the role
of MaybeObject in a more generic way, and indeed MaybeObject becomes a
typedef for MaybeWeak<Object>. The methods on MaybeObject and
HeapObjectReference become free functions, or are removed entirely in
favour of implicit conversions (after all, any strong ref is also a
maybe weak ref).

This requires a few new specializations of Tagged, and some additional
conversion logic. This CL is designed to be minimally invasive,
converting MaybeObject to Tagged<MaybeObject> but trying to use the
typedef and TaggedImpl methods as much as possible to minimise
non-trivial code changes. In the future, we'll want to add a distinct
Weak<T> for definitely weak values, and have conversions from
Tagged<MaybeWeak<T>> to Tagged<T> that don't erase the type.

Bug: v8:12710
Change-Id: I1b5df70b3d2028d46a09ca2cd9bd894a3fd8cb27
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5268603
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92409}