[cppgc] Revise "sweeping out of work" condition
On ToT, out of work means the concurrent sweepers are out of work and
there are no unswept pages left. There could still be unfinalized pages
that need handling on the main thread. This may take a long time and be
too expensive for a minor GC.

This CL revises the condition such that there should also be no
unfinalized pages. This may happen if pages are finalized as part of
on allocation sweeping.

Change-Id: I5e2b08d30a9a8f00c26d6a96a570b33fc165a602
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5285155
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92448}