[turboshaft][wasm] Fix reachability tracking for unreachable loop
This requires a loop that consists of a single block at the time of
running the WasmGCTypeAnalyzer whose predecessor is unreachable.
In that case the loop gets marked as unreachble as well.
If there are some previous outer loops, we can end up revisiting this
unreachable loop. At that point in time we have backedge information
for this loop and we run the MergeSnapshot(block) for this loop after
resetting the reachability of the loop.
The backedge is now marked as reachable and the whole loop will be
marked as unreachable.

This CL fixes it by marking a loop always unreachable if its forward
edge is unreachable.

Ideally, the loop revisitation of previous loops would not revisit
blocks outside of their respective loop bodies.
Unfortunately the AnalyzerIterator isn't smart enough for that.
(It should also be possible to construct a similar scenario where the
unreachable loop is embedded in a reachable outer loop that requires
revisit, in that case we'd probably end up hitting the exact same
DCHECK independently of the AnalyzerIterator visitation order.)

Fixed: 325756545
Change-Id: I7cf4a4d8a51e9c4bc8cf580171bfc2ca73068a14
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5309896
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92429}