[turboshaft][wasm] Fix load type for LOAD_TAGGED_ROOT
The root pointers are stored as uncompressed pointers, therefore
loading from the roots table should load with pointer size.
This breaks on big endian platforms otherwise as we are loading the
wrong half of the pointer then.

This change also simplifies the existing macros to load a root value
with a single macro that uses the root index information to figure out
which load to generate and whether a bitcast is required.

Change-Id: I4b419a7707bf811691dfeebe1e6e89f225dd7870
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5309903
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92439}