[wasm][cleanup] Simplify wasm-to-js wrapper
In the wasm-to-js graph builder, the "arity match" branch is just a
special case of the "arity mismatch" branch. Merge the two cases into
one.
In the "arity match" case, we did not use the {expected_arity} argument.
Use it to check that the arity does match in debug mode, which is
required now for correctness, and fix some places where we forgot to
take JSPI into account.

R=ahaas@chromium.org

Change-Id: I430582215f6ccf9a7f3cd52bd83115bdf4504788
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5309905
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92449}