[heap] MinorMS cycles are never memory reducing
Enforce and document (in DCHECKs) that MinorMS cycles should never be
memory reducing. We expect all memory reducing GCs to be full GCs.

This may improve jank in the rare case of finalizing a concurrent
MinorMS because a full GC is requested. In such cases we currently
finish the minor GC and force finalize sweeping, before we start the
full GC. Sweeping for memory reducing GCs is expensive so force
finalizing it right after it started may actually cause jank.

Bug: v8:12612
Change-Id: If3fbe9f778ac5dff700c7c1d765f5bc15bcbd8c7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5310169
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92451}