[test262] Update renamed tests
On import it can happen that tests got renamed in the test262 while they are still referenced by the old name in the status file. We take care of renames in the status file in this CL (and refactored the test removal case).

No-try: True
Bug: v8:14002
Change-Id: I73ee6e578a1515138e6dc1db365688f64860819b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5309910
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Liviu Rau <liviurau@google.com>
Cr-Commit-Position: refs/heads/main@{#92473}