[sandbox] Ensure safe execution in case of invalid enum in RegExp::Exec
The Type enum value of a JSRegExp is stored as a Smi on the heap and so
it must be assumed that an attacker can change it arbitrarily. In
RegExp::Exec, the enum loaded in this way is used for a `switch`
statement with no default case. In this case, the compiler should
generate code such that an invalid enum value executes none of the
cases. In particular, as the compiler would like to emit a jump table
for this switch, it should also emit a bounds check before that. However, the (non-void) function doesn't return a value after the switch
(only the individual cases do), as such reaching this point is UB and
the compiler can assume that this will never happen. Therefore, the
compiler omits the bounds-check, leading to an OOB access into the jump
table when an invalid enum value is observed.

Simply putting an UNREACHABLE() after the switch fixes this.

Bug: b/326086002
Change-Id: I8ad5d47f90edd081da80e0dc9c1768e70dfb8a54
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5319496
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92504}