Fix iterator incrementation in loop after erase operation
The previous implementation contained a logical error where
the iterator was incremented (++it) in the second branch of
the loop even after the 'erase' operation in the first branch.
This could lead to undefined behavior. The correct approach
is to use 'it = container->erase(it)' to update the iterator
after erasing the element. This ensures proper iterator handling
and avoids potential issues with iterator invalidation.

Bug: none
Change-Id: I7ca307cf5b1a023a39a14aeb867383cbd08cab81
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5320060
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92538}