[sandbox] Fix type confusion in compiler when corrupting meta maps
During compilation, the JSHeapBroker provides access to the heap object
graph, for which it potentially serializes (snapshots) parts of it. Type
checking in that serialized graph happens based on the instance type
obtained from the (serialized) map. However, in case an object's map
points to itself, we assume that this object must be a meta map and so
of type MAP_TYPE. This, however, isn't necessarily true in case of heap
memory corruption, in which case other code may get confused. This CL
addresses that by CHECKing that objects that we believe to be a (meta)
map really are Map objects when serializing the original object.

Bug: b/326700497
Change-Id: If51433143bcd67766d483e3b164aad24ad544cfb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5321525
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92529}