[json] Extract JSDataObjectBuilder code to use an iterator
Extract the fast -> slow path iteration logic in BuildJsonObject to a
helper on JSDataObjectBuilder which iterates an iterator provided by
BuildJsonObject. This moves the complexity of switching between fast and
slow property initialisation into the helper, and the only thing the
iterator has to do is allow re-visiting the values visited in the fast
path.

Change-Id: I8f567d284c8b537d9f9d986128ffba3d2da09289
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5321529
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92564}