[compiler][fastcall][handles] Move AdaptLocalArgument
This CL moves method AdaptLocalArgument so that it is accessible
via the __ macro. This method is used to correctly pass parameters
of type v8::Local<T>; notice that with v8_enable_direct_local=true,
which requires conservative stack scanning, locals are just direct
pointers to objects.

It also merges this method with AdaptHandlifiedArgument, that was
used in the wasm compiler for the same purpose.

Bug: v8:13270
Change-Id: I4ce39eb745dda7fc5b0dba93431d95dd82b9a2ec
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5321466
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92582}