[fuzzer][wasm] Add fuzzer for init expressions
This CL adds a new fuzzer that fuzzes initializer expressions.
Initializers (e.g. for globals, tables, ...) are evaluated using an
interpreter and therefore don't share a common implementation with the
functions that are run via Liftoff Or Turbofan.
The fuzzer creates globals with random initializers and corresponding
functions with the same body as these initializers.

Then, the functions are executed and compared to the initial value of
the global.

Possible next steps:
- Fuzz more expressions in the initializers (e.g. `global.get`)
- Use initializers not only in globals but also for Tables etc.?

Change-Id: I76ea8aabe9792c5c749da87a7b49277e37acb136
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5324481
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92570}