[sandbox] Fix unsafe switch in FeedbackMetadata::GetSlotSize
This is similar to b/326086002: we have an exhaustive switch over an
enum with no default case and are missing a `return` or `UNREACHABLE`
after the switch, causing the compiler to generate unsafe code in case
an invalid enum value is used, which can happen in the sandbox attacker
model. The fix is the same as for b/326086002: add the `UNREACHABLE`.

Bug: b/326086002
Change-Id: I9455b9e36850e39e9635ca7f6555487c40e8626a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5328109
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92566}