[wasm] Avoid storing the whole signature for fast API calls
Booleans in C++ are a single bytes, and if loaded into a register, the
upper bytes in the register may be non-zero. When the C++ bool gets
converted into a WebAssembly I32, the values in the upper bytes of the
register therefore become observable if no explict zero-extension is
happening in WebAssembly. Therefore WebAssembly has to know if the
return value of a fast API function is boolean, so that the result value
gets converted.

So far the whole signature of the fast API function was provided to the
wasm compiler, which included the information if the return value was
boolean. However, 1) the whole signature does not seem to be needed, and
2) the signature may get deallocated before it gets accessed by the
compilation. This can happen because each isolate has its own copy of
the signature, but the wasm compilers don't have access to the signature
of the isolate, only to a pointer to the signature of the last
registered isolate in the NativeModule. Therefore, if the last
registered isolate dies before the wasm compiler accesses the signature,
then the wasm compiler accesses a stale pointer.

Therefore the NativeModule now only stores the information if the return
type is bool.

Additionally the array in the NativeModule that stores if the return
type is bool is changed to std::atomics<bool> to avoid a TSAN report.

R=evih@chromium.org

Bug: v8:14638
Change-Id: I7ae1b9c0a088a9f12c4c127d1189cd4537cbd3ac
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5331712
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92589}