[d8][inspector][test] Fix pumping the message loop
When pumping the message loop, we may trigger the execution of a
non-nestable task (e.g., a stackless GC), which assumes that there are
no handles on the current stack. With v8_enable_direct_local=true
(which requires conservative stack scanning), violating this invariant
may result in a crash. This CL fixes d8 and the inspector task runner,
installing a scope to enforce scanning the stack in this case. It also
fixes a unit test that was occasionally failing with conservative stack
scanning.

Bug: v8:13270
Change-Id: I48916866b7bd4ab235736073b66827e71d78feb8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5331539
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92599}