Make Continuation Preserved Embedder Data more robust
This CL fixes 3 issues with CPED found while simplifying blink's
task attribution (and decreasing how frequently we propagate state).

Note (A): When running a promise reaction job, the CPED stored on the
reaction object is only restored if it's not undefined (done as an
optimization). If CPED was modified prior to running a microtask, it's
set to undefined after.

 1. From (A), CPED can never be set to Undefined, otherwise the CPED
    can be incorrect (it can leak the previous value). This isn't a
    problem for blink's Task Attribution currently since it ensures
    there's something to propagate for each JS callback, but we're
    working towards propagating less frequently, in which case Undefined
    will be valid / be the default when there's nothing to propagate.

 2. Also from (A), since the previous CPED value isn't restored after a
    microtask runs, CPED can become undefined when the embedder might
    expect it to still be set. In blink, this happens in event
    processing; for example, if CPED/task state is set before
    dispatching an event, it can become undefined between callbacks if
    there are multiple event listeners.

 3. Only promise reactions propagate CPED, but blink makes extensive
    use of the EnqueueMicrotask API. There was a hack in place in blink
    to fall back to a cached version of CPED, but that has edge cases
    (CPED and the cached version get out of sync when cleared in (A)).

To fix (1) and (2), this CL caches the the Isolate's CPED before running
microtasks and clears/restores it in RunMicrotasks. This leaves the
optimization in place to only set the CPED if it's not undefined since
it's being cleared at the start of the checkpoint and after each
microtask.

To fix (3), this CL adds the isolate CPED as a field in Microtask and
makes all Microtask subclasses CPED-aware (new for EnqueueMicrotask and
PromiseResolveThenableJobTask tasks). Blink _probably_ only needs the
former, but being consistent seemed better here.

We could alternatively do these in embedder, but at least (1) seems like
something the API should support (unless we require embedders to clear
CPED before a microtask checkpoint). (2) can be solved in the embedder
by storing/restoring CPED after each callback, but it's simpler/cheaper
to do in v8 and seems like a reasonable API guarantee. (3) could also be
done in the embedder for EnqueueMicrotask, but it would also be more
expensive (likely involving an GC allocation to wrap the isolate's
CPED value).

Bug: b/326368637
Change-Id: If6b93c68146f0d8c01a59b4d37e869ec25829a9d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5316657
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92616}