[sandbox] Fix stack out-of-bounds write when heap memory is corrupted
In Intl::ToICUUnicodeString we have a fast path for small strings where
the contents is copied into an 80-byte large stack-allocated buffer.
However, the check to determine whether the string was small enough was
performed on signed integers. As an attacker could corrupt the on-heap
string length to be a negative number, that check would pass and the
code would continue with essentially a wild memcpy onto the stack.

Bug: b/327473424
Change-Id: I24c0fb7b7b2b40b24bf8a129a17c9bb8a1904d5b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5335156
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92601}