[static-roots] Fix fast setting of a function callback return values
This CL ensures that the upper 32-bits of the ReturnValue<T> are never
ever modified by making SetInternal() function a bottleneck and using
it everywhere.

The issue was that setting a Smi value corrupted the upper 32-bits
and subsequent setting of an undefined/null/true/false didn't
recover them.

Drive-by: use SetNonEmpty() for relevant use cases.

Bug: v8:14445
Change-Id: I7f3da1469afed2f5c23e670e13ef89fdd7bf573a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5331689
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92621}