[sandbox] Defend against corrupted Map in JSNativeContextSpecialization
Normally, map.NextFreePropertyIndex() will always be larger than
map.GetInObjectProperties(). However, an attacker able to corrupt heap
memory can break this invariant by corrupting a Map object. In that
case, BuildExtendPropertiesBackingStore in JSNativeContextSpecialization
would get confused as it subtracts the two and expects a positive
integer as result, which is then used as length of a vector. To defend
against this, this CL simply inserts a SBXCHECK.

This bug is quite similar to b/326355031.

Bug: b/327518873
Change-Id: I3b0aba22c26cffa7a05cffec0ddd8346dc938740
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5335696
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92619}