[sandbox] Do not use SA_RESETHAND when installing the crash filter
This flag causes the handler to be reset to the default handler upon
invocation, which is probably not what we want: for example, it can
happen that corrupting heap memory causes two threads to segfault in
short succession. In that case, the first thread will invoke the crash
filter, but the second may cause the process to terminate with the
SEGFAULT before the crash filter has finished executing. Instead, we
probably want both threads to execute the crash filter in this case.

Bug: v8:12878
Change-Id: I3971c621db216b5b2782004908e9dcfe584604df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5331699
Auto-Submit: Samuel Groß <saelo@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#92633}