[heap][handles] Stack scanning for background threads
This CL adds support for conservatively scanning the stacks of
background threads. If such a thread has a non-trivial stack, it
should use Stack::SetMarkerForBackgroundThreadAndCallback to register
its stack for scanning and wait in the callback for a GC to occur.
This method is automatically invoked by the family of
LocalHeap::BlockWhileParked methods.  Threading cctests use the
same mechanism for alternating between different threads.

Bug: v8:13257
Bug: v8:13270
Change-Id: I9bcced9e80d2e07f2a4a02f45cb6d129a79a16ce
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5331543
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92629}