[heap] Embed the MemoryChunk in MemoryChunkMetadata
instead of using a subclass.
This enforces that all Metadata => Chunk accesses go through the right
accessors and the other way round.
This is a no-op at the moment, but will be needed once the Metadata
moves out of the Chunk.
In a follow up, I'll do some refactoring to remove unnecessary lookups.

Change-Id: I91e8bce5dcba611b87738237522a7031ae432960
Bug: v8:13355 chromium:40849120
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5339895
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#92655}