[turbofan] Move protected pointer loads to CodeAssembler
This refactors the API to be in line with default on-heap field loads:
The CodeStubAssembler works with field offsets and calls out to a
lower-level method in the CodeAssembler which works with raw offsets.
The naming now also follows that of other field loads.

R=dmercadier@chromium.org

Bug: v8:14564
Change-Id: I7289d1ae3443e2a142b7d3515f834da28537af2c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5331725
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92669}